+++ src/auth/components/LoginPage/LoginPage.tsx
@@ -10,23 +10,20 @@ import Typography from "@material-ui/core/Typography";
 import React from "react";
 import { FormattedMessage, useIntl } from "react-intl";
 
-import { ControlledCheckbox } from "@saleor/components/ControlledCheckbox";
 import Form from "@saleor/components/Form";
 import { FormSpacer } from "@saleor/components/FormSpacer";
 import { commonMessages } from "@saleor/intl";
 
 export interface FormData {
   email: string;
-  loading: boolean;
   password: string;
-  rememberMe: boolean;
 }
 
 const styles = (theme: Theme) =>
   createStyles({
     buttonContainer: {
       display: "flex",
-      justifyContent: "space-between"
+      justifyContent: "flex-end"
     },
     link: {
       color: theme.palette.primary.main,
@@ -65,10 +62,7 @@ const LoginCard = withStyles(styles, { name: "LoginCard" })(
     const intl = useIntl();
 
     return (
-      <Form
-        initial={{ email: "", password: "", rememberMe: false }}
-        onSubmit={onSubmit}
-      >
+      <Form initial={{ email: "", password: "" }} onSubmit={onSubmit}>
         {({ change: handleChange, data, submit: handleSubmit }) => (
           <>
             {error && (
@@ -93,7 +87,7 @@ const LoginCard = withStyles(styles, { name: "LoginCard" })(
             <FormSpacer />
             <TextField
               fullWidth
-              autoComplete="current-password"
+              autoComplete="password"
               label={intl.formatMessage({
                 defaultMessage: "Password"
               })}
@@ -101,20 +95,12 @@ const LoginCard = withStyles(styles, { name: "LoginCard" })(
               onChange={handleChange}
               type="password"
               value={data.password}
-              data-tc="password"
+              inputProps={{
+                "data-tc": "password"
+              }}
             />
             <FormSpacer />
             <div className={classes.buttonContainer}>
-              <ControlledCheckbox
-                checked={data.rememberMe}
-                label={intl.formatMessage({
-                  defaultMessage: "Remember me",
-                  description: "login"
-                })}
-                name="rememberMe"
-                onChange={handleChange}
-              />
-              <FormSpacer />
               <Button
                 className={classes.loginButton}
                 color="primary"